Traits implementados. Modificada la implementacion de Cancha.

+++ tp1-mixins-complejodeportivo/src/canchas/Futbol.scala
@@ -2,6 +2,6 @@ package canchas
 
 class Futbol(var cantidadDeJugadores : Int) extends Cancha {
 
-  def precio() : Int = 40 * this.cantidadDeJugadores
+  def precio() : Double = 40 * this.cantidadDeJugadores
 }
